productEdit not fixed yet

+++ database/migrations/2019_06_06_190630_create_cake_sizes_table.php
@@ -15,7 +15,7 @@ class CreateCakeSizesTable extends Migration
     {
         Schema::create('cake_sizes', function (Blueprint $table) {
             $table->integer('id_cake');	
-            $table->integer('sizes');	
+            $table->float('sizes');	
         });
     }
